refactor some util logic into own header
